worker/requirements.txt
flask>=2.0
ping3>=3.0
dnspython>=2.0
rdap>=1.2